+++ src/modules/post.jsx
@@ -1,18 +1,12 @@
 import { createAction, handleActions } from "redux-actions";
-import createRequestSaga, { createRequestActionTypes } from "../lib/createRequestSaga";
+import createRequestThunk, { createRequestActionTypes } from "../lib/createRequestThunk";
 import * as postAPI from '../lib/api/posts';
-import { takeLatest } from "redux-saga/effects";
 
 const [READ_POST,READ_POST_SUCCESS,READ_POST_FAILURE]=createRequestActionTypes('post/READ_POST');
 const UNLOAD_POST='post/UNLOAD_POST';
 
-export const readPost=createAction(READ_POST,id=>id);
+export const readPost=createRequestThunk(READ_POST,postAPI.readPost);
 export const unloadPost=createAction(UNLOAD_POST);
-
-const readPostSaga=createRequestSaga(READ_POST,postAPI.readPost);
-export function* postSaga(){
-    yield takeLatest(READ_POST,readPostSaga);
-}
 
 const initialState={
     post:null,
