+++ src/modules/user.jsx
@@ -1,39 +1,14 @@
 import { createAction, handleActions } from "redux-actions";
-import createRequestSaga, { createRequestActionTypes } from "../lib/createRequestSaga";
+import createRequestThunk, { createRequestActionTypes } from "../lib/createRequestThunk";
 import * as authAPI from '../lib/api/auth';
-import { takeLatest, call } from "redux-saga/effects";
 
 const TEMP_SET_USER='user/TEMP_SET_USER';
 const [CHECK,CHECK_SUCCESS,CHECK_FAILURE]=createRequestActionTypes('user/CHECK');
 const LOGOUT='user/LOGOUT';
 
 export const tempSetUser=createAction(TEMP_SET_USER,user=>user);
-export const check=createAction(CHECK);
-export const logout=createAction(LOGOUT);
-
-const checkSaga=createRequestSaga(CHECK,authAPI.check);
-
-function checkFailureSaga(){
-    try{
-        localStorage.removeItem('user');
-    }catch(e){
-        console.log('localstorage error');
-    }
-}
-function* logoutSaga(){
-    try{
-        yield call(authAPI.logout);
-        localStorage.removeItem('user');
-    }catch(e){
-        console.log(e);
-    }
-}
-
-export function* userSaga(){
-    yield takeLatest(CHECK,checkSaga);
-    yield takeLatest(CHECK_FAILURE,checkFailureSaga);
-    yield takeLatest(LOGOUT,logoutSaga);
-}
+export const check=createRequestThunk(CHECK,authAPI.check);
+export const logout=createAction(LOGOUT,authAPI.logout);
 
 const initialState={
     user:null,
